Front and backend merge
I have now merged the frontend and the backend so that now it all works together. In the public/index.php I have included a very simple ajax call that will update the users table automatically.

As you can see from the ajax call, it is using the `/api/user` API call. This is good because it relying on the API. This means that third party apps, plugins and scripts can all use the same data.