Added Bearer Token Authentication and Improved Security and Performance
In this commit, we have implemented a secure and reliable Bearer token generator, with advanced validation and verification methods to enhance the app's security. We have also optimized the namespaces for the Streets and User controllers, ensuring a clean and organized code structure. Moreover, we have updated the .htaccess file with additional security measures to prevent any unauthorized access or potential security breaches. This commit marks a significant milestone in improving the overall security and performance of the app, ensuring that it provides a secure and reliable user experience.

In addition to these changes, we have also made some improvements to the database connection by utilizing the Database class to make the connection to the database more secure and efficient. We have updated the AuthController to encrypt the user's passwords with the default PHP password hash method, ensuring an additional layer of security. These updates not only enhance the app's security but also optimize its performance and functionality. Overall, this commit represents a significant step in improving the app's security, usability, and performance, making it more reliable and secure for our users.